requirements.txt
Flask>=2.0
requests>=2.28
python-dotenv>=1.0

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Document Loaders
PyPDF2>=3.0
python-docx>=1.0

# Vector Store
faiss-cpu>=1.7.0

# Database
sqlalchemy>=2.0
